added auth views

+++ base/views.py
@@ -1,14 +1,60 @@
 # views.py
 
 from django.shortcuts import render, get_object_or_404, redirect
-from .models import Room
+from django.http import HttpResponse
+from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
+from .models import Room, Topic
 from .forms import RoomForm
 
 # Create your views here.
 
+def loginPage(request):
+
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User doesnt exist' )
+
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Invalid Credentials')
+
+
+
+    context={'page':page}
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
 def home(request):
-    rooms = Room.objects.all()
-    context = {'rooms': rooms}
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+    rooms = Room.objects.filter(Q(topic__name__icontains=q) |
+                                Q(name__icontains=q) |                                
+                                Q(description__icontains=q)
+              )
+
+    topics = Topic.objects.all()
+    room_count = rooms.count()
+    context = {'rooms': rooms, 'topics':topics, 'room_count':room_count}
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
@@ -16,8 +62,13 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
+
+    if request.user != room.host:
+        return HttpResponse('You are not Allowed here here !!')
+    
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -27,9 +78,14 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not Allowed here here !!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -40,8 +96,14 @@ def updateRoom(request, pk):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not Allowed here here !!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
